01_Generador_Imagenes_Streamlit/requirements.txt
streamlit>=1.31.0
openai>=1.30.0,<2.0.0
Pillow>=10.0.0
